Cleaned up code
Cleaned up code
