+++ public/index.html
@@ -42,7 +42,7 @@
       rel="stylesheet"
     />
 
-    <title>React App</title>
+    <title>Romantic Comedy</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
